Удалены ненужные и неактуальные конструкторы, и методы. Подготовлен родительский класс к дальнейшему выполнению ТЗ.
