+++ src/transport/Transport.java
@@ -4,93 +4,40 @@ public abstract class Transport {
 
     private final String brand;
     private final String model;
-    private final int year;
-    private final String country;
-    private String color;
-    private int maxSpeed;
+    private double engineVolume;
 
-    public Transport(String brand, String model, int year, String country, String color, int maxSpeed) {
-        if (brand == null || brand.isBlank() || brand.isEmpty()) {
-            this.brand = "default";
-        } else {
-            this.brand = brand;
-        }
-
-        if (model == null || model.isBlank() || model.isEmpty()) {
-            this.model = "default";
-        } else {
-            this.model = model;
-        }
-
-        if (year > 0) {
-            this.year = year;
-        } else {
-            this.year = 2000;
-        }
-
-        if (country == null || country.isBlank() || country.isEmpty()) {
-            this.country = "default";
-        } else {
-            this.country = country;
-        }
-
-        setColor(color);
-        setMaxSpeed(maxSpeed);
+    public Transport(String brand, String model, int engineVolume) {
+        this.brand = brand;
+        this.model = model;
+        this.engineVolume = engineVolume;
     }
 
-    public Transport(String brand, String model, int year, String country) {
-        this(brand, model, year, country, "white", 120);
-    }
+    public String getBrand() {
 
-    public String getColor() {
-        return color;
-    }
-
-    public final void setColor(String color) {
-        if (color == null || color.isBlank() || color.isEmpty()) {
-            this.color = "white";
-        } else {
-            this.color = color;
-        }
-    }
-
-    public int getMaxSpeed() {
-        return maxSpeed;
-    }
-
-    public final void setMaxSpeed(int maxSpeed) {
-        if (maxSpeed <= 0) {
-            this.maxSpeed = 120;
-        } else {
-            this.maxSpeed = maxSpeed;
-        }
-    }
-
-    public final String getBrand() {
         return brand;
     }
 
-    public final String getModel() {
+    public String getModel() {
         return model;
     }
 
-    public final int getYear() {
-        return year;
+    public double getEngineVolume() {
+        return engineVolume;
     }
 
-    public final String getCountry() {
-        return country;
+    public void setEngineVolume(int engineVolume) {
+        if (engineVolume <= 0) {
+            this.engineVolume = 1.5;
+        } else {
+            this.engineVolume = engineVolume;
+        }
     }
 
-    @Override
-    public String toString() {
-        return "Transport{" +
-                "brand='" + brand + '\'' +
-                ", model='" + model + '\'' +
-                ", year=" + year +
-                ", country='" + country + '\'' +
-                ", color='" + color + '\'' +
-                ", maxSpeed=" + maxSpeed +
-                '}';
+    public void startMoving() {
+        System.out.println("Driving");
+    }
+
+    public void endMoving() {
+        System.out.println("Stopped");
     }
 }
