+++ src/transport/Truck.java
@@ -1,0 +1,7 @@
+package transport;
+
+public class Truck extends Transport {
+    public Truck(String brand, String model, int engineVolume) {
+        super(brand, model, engineVolume);
+    }
+}
